[Planificador] Comando block

+++ planifier/orchestrator.c
@@ -63,7 +63,7 @@ bool is_valid_esi(long esi_id){
 
 
 
-void block_esi(int esi_id){
+void block_esi(long esi_id){
 	modificar_estado(esi_id, BLOQUEADO);
 	switch(ALGORITHM) {
 			case FIFO:
